Split cli.py => {commands.py, core.py}
This makes much more sense, for now. Remove prefix _command
for commands and use namespace commands.function to give meaning.

BTW, all the usage of api.py need be here on mal.core
