+++ .env
@@ -1,0 +1,1 @@
+source env/bin/activate
