+++ mal/utils.py
@@ -8,7 +8,6 @@
 #
 
 import sys
-import mal
 import os
 from functools import wraps
 from sre_constants import error as BadRegexError
@@ -39,31 +38,6 @@ class StopSpinner(object):
 # global variables from hell
 sys.stdout = Unbuffered(sys.stdout)
 sig = StopSpinner()
-
-
-def usage():
-    usage_info = [
-        (("Usage: mal [inc/+1 | dec/-1] anime-by-regex\n"
-          "       mal [watching | plan to watch | ... | rewatching]\n"
-          "       mal [list | all]\n"
-          "       mal anime-by-regex\n")),
-        (("Ex. for increment +1:\n\n\t"
-          "$ mal +1 lain\n")),
-        (("Ex. for decrement -1:\n\n\t"
-          "$ mal -1 lain\n")),
-        (("Ex. filtering for section:\n\n\t"
-          "$ mal watching\n")),
-        (("Ex. search return all anime whose start with s: \n\n\t"
-          "$ mal ^s\n")),
-        (("Ex. fetch all list: \n\n\t"
-          "$ mal list\n\t"
-          "$ mal all\n\t"
-          "$ mal .+\n")),
-        "Hacked by {__author__} <{__email__}> | "
-        "version {__version__}".format_map(vars(mal)),
-    ]
-    print('\n'.join(usage_info), file=sys.stderr)
-    sys.exit(1)
 
 
 def killed():
